requirements.txt
# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# Google Gemini AI
google-generativeai>=0.8.0

# LangChain & RAG
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.1.0

# Vector Database
chromadb>=0.5.0

# Document Processing
docx2txt>=0.8

# Environment Variables
python-dotenv>=1.0.0
